move assert to separate util file
